Update custom Dask client and docs
Much of the hackery in the custom Dask client to workaround Dask's
limitations with serializing proxies are no longer needed with the
changed in ProxyStore v0.6.5. Now the custom Dask client's purpose is
just automatic proxying of objects greater than a threshold.

The Dask guide has been updated to just point users to other more
up-to-date sources of information.