Wrap proxied arguments in dask DataNode type
Dask distributed recently (PR 8797) changes the representation of tasks
from a tuple of (func, args, kwargs) to a Task. The Task types parses
the args and kwargs to tokenize into GraphNode objects. This parsing
function (dask._task_spec.parse_input:
https://github.com/dask/dask/blob/2024.11.1/dask/_task_spec.py#L116-L146)
performs some isinstance checks for dict, list, tuple, set, etc. which
are true for proxies of those types. However, the function then tries
to recursively call itself on each element of that collection,
recreating a new collection in the process. This does not work with a
proxy type and causes an error.

To address this, the argument proxying functions here have been modified
to return GraphNode instances which are ignored by parse_input.
Specifically, a new _ProxyNode that is inherits from DataNode is use.
The data of the DataNode is just the proxy, which (as far as I can tell)
preserves the old behavior with Dask.